requirements.txt
flask==2.3.2
flask_httpauth==4.8.0
PyJWT==2.8.0